会员成就

+++ src/main/java/com/training/entity/MemberCouponEntity.java
@@ -10,7 +10,7 @@ import java.util.Date;
 @Data
 public class MemberCouponEntity {
 
-    private Long couponId;
+    private Integer couponId;
 
     private String memberId;
